Detaches FSM Actor logic to allow overriding
