+++ src/main/scala/com/suprnation/actor/fsm/FSM.scala
@@ -209,203 +209,22 @@ case class FSMBuilder[F[+_]: Parallel: Async: Temporal, S, D, Request, Response]
       onTerminationCallback = Option(onTerminated)
     )
 
-  final def startWith(stateName: S, stateData: D, timeout: Timeout = None): PreCompiledFSM = {
+  final def startWith(
+      stateName: S,
+      stateData: D,
+      timeout: Timeout = None,
+      actorBuilder: FSMActor.Constructor[F, S, D, Request, Response] =
+        FSMActor.defaultConstructor[F, S, D, Request, Response]
+  ): PreCompiledFSM =
     new PreCompiledFSM { preCompiledSelf =>
-      override def initialize: F[ReplyingActor[F, Request, Response]] = for {
-        currentStateRef <- Ref.of[F, State[S, D, Request, Response]](
-          State(stateName, stateData, timeout)
-        )
-        nextStateRef <- Ref.of[F, Option[State[S, D, Request, Response]]](None)
-        generationRef <- Ref.of[F, Int](0)
-        timeoutFiberRef <- Ref.of[F, Option[Fiber[F, Throwable, Unit]]](None)
-      } yield {
-        (new Actor[F, Any] with ActorLogging[F, Any] {
-          // State manager will allow us to move from state to state.
-          val stateManager: StateManager[F, S, D, Request, Response] =
-            new StateManager[F, S, D, Request, Response] {
-              override def goto(nextStateName: S): F[State[S, D, Request, Response]] =
-                currentStateRef.get.map(currentState =>
-                  State(nextStateName, currentState.stateData)
-                )
-
-              override def stay(): F[State[S, D, Request, Response]] =
-                (currentStateRef.get >>= (currentState => goto(currentState.stateName)))
-                  .withNotification(false)
-
-              override def stop(reason: Reason): F[State[S, D, Request, Response]] =
-                currentStateRef.get >>= (currentState => stop(Normal, currentState.stateData))
-
-              override def stop(reason: Reason, stateData: D): F[State[S, D, Request, Response]] =
-                stay().using(stateData).withStopReason(reason)
-
-              override def forMax(
-                  finiteDuration: Option[(FiniteDuration, Request)]
-              ): F[State[S, D, Request, Response]] =
-                stay().map(_.forMax(finiteDuration))
-
-              override def stayAndReply(replyValue: Response): F[State[S, D, Request, Response]] =
-                stay().map(_.replying(replyValue))
-            }
-
-          private def cancelTimeout: F[Unit] =
-            timeoutFiberRef.get.flatMap(_.fold(Sync[F].unit)(_.cancel))
-
-          private def handleTransition(prev: S, next: S): F[Unit] =
-            transitionEvent.collect {
-              case te if te.isDefinedAt((prev, next)) => te((prev, next))
-            }.sequence_
-
-          private val handleEvent: StateFunction[Any] = { case (Event(value, _), _) =>
-            for {
-              currentState <- currentStateRef.get
-              _ <- log(
-                s"[Warning] unhandled [Event $value] [State: ${currentState.stateName}] [Data: ${currentState.stateData}]"
-              )
-            } yield currentState
-          }
-
-          override def preStart: F[Unit] = currentStateRef.get >>= makeTransition
-
-          override def receive: ReplyingReceive[F, Any, Any] = {
-            case TimeoutMarker(gen, sender, msg) =>
-              generationRef.get
-                .map(_ == gen)
-                .ifM(
-                  processMsg(StateTimeoutWithSender(sender, msg), "state timeout"),
-                  Sync[F].unit
-                )
-
-            case value =>
-              cancelTimeout >> generationRef.update(_ + 1) >> processMsg(value, sender)
-          }
-
-          private def processMsg(value: Any, source: AnyRef): F[Any] =
-            currentStateRef.get >>= (currentState =>
-              (if (config.debug) config.receive(value, sender, currentState) else Sync[F].unit)
-                >> processEvent(Event(value, currentState.stateData), source)
-            )
-
-          private def processEvent(event: FSM.Event[D, Any], @unused source: AnyRef): F[Any] =
-            for {
-              currentState <- currentStateRef.get
-              stateFunc = stateFunctions(currentState.stateName)
-              // If this is a Timeout event we will set the original sender on the cell so that the user can reply to the original sender.
-              state <- event.event match {
-                case StateTimeoutWithSender(sender, msg) =>
-                  /* The StateTimeoutWithSender was only used to propagate the sender information, the actual user does not need to understand these
-                   * semantics. */
-                  // We will simplify the message to the SenderTimeout, the sender will be implicit on the cell.
-                  val updatedEvent: FSM.Event[D, Request] =
-                    event.copy(event = msg.asInstanceOf[Request])
-                  /* Override the sender to be the same sender as the original state (we might want to notify him that something happened in the
-                   * timeout) */
-                  (self.cell >>= (c =>
-                    Async[F]
-                      .delay(c.creationContext.senderOp =
-                        sender.asInstanceOf[Option[ActorRef[F, Nothing]]]
-                      )
-                  )) >>
-                    stateFunc
-                      .lift((updatedEvent, stateManager))
-                      .getOrElse(
-                        handleEvent((updatedEvent.asInstanceOf[FSM.Event[D, Any]], stateManager))
-                      )
-                case _ =>
-                  stateFunc
-                    .lift((event.asInstanceOf[Event[D, Request]], stateManager))
-                    .getOrElse(handleEvent((event, stateManager)))
-              }
-
-              _ <- applyState(state)
-            } yield ()
-
-          private def applyState(nextState: State[S, D, Request, Response]): F[Unit] =
-            nextState.stopReason match {
-              case None =>
-                makeTransition(nextState)
-
-              case Some(_) =>
-                val sendReplies: F[Unit] =
-                  nextState.replies.reverse.traverse_(r => sender.get.widenRequest[Response] ! r)
-                sendReplies >> terminate(nextState) >> context.stop(self)
-            }
-
-          def makeTransition(nextState: State[S, D, Request, Response]): F[Unit] = {
-            def scheduleTimeout(timeoutData: (FiniteDuration, Request)): F[Unit] =
-              cancelTimeout >>
-                (for {
-                  generation <- generationRef.get
-                  currentSender = sender
-                  cancelFn <- context.system.scheduler.scheduleOnce_(timeoutData._1)(
-                    self !* TimeoutMarker[F, Request](generation, currentSender, timeoutData._2)
-                  )
-                  _ <- timeoutFiberRef.set(Some(cancelFn))
-                } yield ())
-
-            def processStateTransition(currentState: State[S, D, Request, Response]): F[Unit] =
-              (if (currentState.stateName != nextState.stateName || nextState.notifies) {
-                 nextStateRef.set(Some(nextState)) >>
-                   handleTransition(currentState.stateName, nextState.stateName) >>
-                   nextStateRef.set(None)
-               } else Sync[F].unit) >>
-                (if (config.debug) config.transition(currentState, nextState) else Sync[F].unit) >>
-                currentStateRef.set(nextState) >>
-                (nextState.timeout match {
-                  case Some((d: FiniteDuration, msg)) if d.length >= 0 =>
-                    scheduleTimeout(d -> msg)
-                  case _ =>
-                    stateTimeouts(nextState.stateName).fold(Sync[F].unit)(scheduleTimeout)
-                })
-
-            stateFunctions.get(nextState.stateName) match {
-              case None =>
-                val errorMessage = s"Next state ${nextState.stateName} does not exist"
-                stateManager.stay().withStopReason(Failure(errorMessage)) >>= terminate
-
-              case Some(_) =>
-                // If the sender is not not specified (e.g. if you are running directly from an application) we will ignore it to be safe.
-                nextState.replies.reverse.traverse_(r =>
-                  sender.fold(Sync[F].unit)(sender =>
-                    {
-                      // The sender knew that we could send him the Response so this is totally safe.
-                      sender.widenRequest[Response] ! r
-                    }.void
-                  )
-                ) >>
-                  currentStateRef.get.flatMap(processStateTransition)
-            }
-
-          }
-
-          protected def onTerminated(reason: Reason): F[Unit] =
-            onTerminationCallback.fold(Sync[F].unit)(x => x(reason))
-
-          override def postStop: F[Unit] =
-            (stateManager.stay().withStopReason(Shutdown) >>= terminate) >> super.postStop
-
-          private def terminate(nextState: State[S, D, Request, Response]): F[Unit] =
-            for {
-              currentState <- currentStateRef.get
-              _ <- currentState.stopReason.fold {
-                nextState.stopReason.fold(Sync[F].unit) { reason =>
-                  for {
-                    _ <- cancelTimeout
-                    _ <- onTerminated(reason)
-                    _ <- currentStateRef.set(nextState)
-                    stopEvent = StopEvent(reason, currentState.stateName, currentState.stateData)
-                    _ <- terminateEvent.lift(stopEvent).getOrElse(Sync[F].unit)
-                  } yield ()
-                }
-              }(_ => Sync[F].unit)
-            } yield ()
-        }).asInstanceOf[ReplyingActor[F, Request, Response]]
-      }
+      override def initialize: F[ReplyingActor[F, Request, Response]] =
+        (FSMActorState.initialize(stateName, stateData, timeout) >>= {
+          state: FSMActorState[F, S, D, Request, Response] => actorBuilder(builderSelf, state)
+        }).map(_.asInstanceOf[ReplyingActor[F, Request, Response]])
     }
-  }
 
   trait PreCompiledFSM {
 
     def initialize: F[ReplyingActor[F, Request, Response]]
   }
-
 }
